added a cmake build system
This cmake system is expected to replace hibridon's legacy build system to make the build system simpler (standard makefile with clean build and install targets).

In the process, we also replace hibridon's custom preprocessing system with fortrans's standard preprocessing. We do this because this also makes the build process simpler and more standard.

At the moment the cmake build system is far from complete but this is a proof of concept : we can yet build one potentials' executable on linux with gfortran

What still needs doing:
- remove all uses of custom preprocessing
- replace xdate in hiversion.t
- implement make test
- ensure it works for ifort and darwin
- etc..