replaced hibridon's custom code enabling directives with standard fpp directives
As cmake build system uses standard fpp preprocessing, we had to replace hibridon's custom preprocessing directives with fpp directives to make preprocessing work with cmake build system (well, we could have made cmake use hibridon's custom preprocessing, but one of the goals of this refactoring is also to make the code more standard and simpler (the custom preprocessing adds unnecessary complexity)).

In this commit, I replaced things like "cstart unix-hp" with #if defined(HIB_UNIX_HP), preserving all the existing architectures (however, I had to replace `-` with `_` because `-`  is not allowed in symbol names). This was performed using semi automatic search and replace wthin vscode. I commit it even if there are still things to do to make the code buildable