fixed bug that caused the makehib -b flag to have no effect
This bug was discovered while attempting to add hibb_arn2 test to cmake as part of issue #8

Well there is a bug in which the script makehib is not compatible with himain.t (now renamed as himain.F for cmake build). In himain.t, makehib expected the size of tq* matrices to be kbig,kbig, but this was no longer the case: someone set the size of tq* matrices to be kmax,kmax, and this caused makehib -b to no longer work as expected. As a result, the -b option had no effect (mostly ignored)

This commit makes it explicitely clear that the size of tq* matrices are compile time parameters, to prevent accidental regressions like this, even if makehib system is broken (it's being replaced with cmake)