merged autotest's tests (autotest/*) into the test directory (test/*)
before this commit, we had 2 test suites :
- `testsuite1` in tests/* :this test suite was used by the original test tool `hibtest`
- `testsuite2` in autotests/* :this test suite is used by the second test tool (`autotest`), and has been build from `testsuite1`, then expanded with other tests. As a result, `testsuite2` duplicates the `testsuite1` tests.

We also have 3 test tools:
- `hibtest` : hardcoded script
- `autotest` : written in python, improvement on `hibtest`:
   - provides automatic results file comparison for most result files
   - modular design : one directory for each test
   - no need to modify code to add a test
   - still uses hibridon's original build system `makehib`, and therefore is not compatible with ctest.
- `ctest` :  test system provided by cmake, and provides nearly the same features as `autotest` (however, one can argue that we still have to write script to add a new test: this is true but this commit makes this script as simple as it can)

This commit merges `testsuite2` and `testsuite1` into `testsuite3` which:
   - resides in `tests/*`
   - is in a form compatible with the `autotest` tool and `ctest` (but no longer compatile with `hibtest`) : one directory for each test is much cleaner.

This commit also improves the cmake test scripts to make them as easy to to as autotest tests.

note: sometimes it was tricky to decide if a test in testsuite1 was redundant from a test in testsuite2; I hope I didn't make mistakes.

This work is related to issue #19. The next step is to enable all tests in the now expanded testsuite3.

In the future, it is planned to remove autotest tool (even if it was well designed) as:
1. it's not compatible with cmake, which brings lots of benefits compated to hibridon's legacy build system.
2. ctest does everything autotest does, and it's better to use a standard system (less code to maintain)