added option flags to help detect uninitialized variables when compiling with gfortran
Without these flags gfortran initializes variables to a value of zero, which is probably suitable for the uninitilaized variables that hibridon contains. However, there is no option in ifort to have the same behaviour, as these are actual errors that need to be fixed.

These options are disabled yet because the cause lots of tests to fail, and this explains why lost of tests fail with ifort too... It is planned to enable these gfortran options once all tests pass with them.

work related to issue #37