converted fixed form fortran code into free form
Free form will allow the assert system to work, as it needs unlimited line length

This conversion was done with the help of the python module fixed2free2.py, then hand tweaking

note: I didn't change the file extension yet because if I did, git would lose the history of the files. The renaming will be performed in a second step. As a result, this commit is broken.