documented the recommended way to build user defined potentials
The old way (through the use of user.cmake in build dir)  had the problem that it is messy to put source code (such as user.cmake) in a directory that is prone to deletion. True, we could have chosen to put these user.cmake files in hibridoun's source directory but this is not ideal either because it messes with hibridon's source tree.

This new way is more natural; it treats hibridon as what it is: a library; and the user creates his own project that uses this library.