refactored `hiba01_1sg` basis into a class
From now on, the bases are implemented as classes, but only `hiba01_1sg` has been refactored to fit this refactoring. And therefore only the test `heco` which uses this base has been enabled.

The other bases will be refactored too when ibasty is no longer used in the code (which will mean that the process of converting basis into a class abstraction is complete)

Work done with @ppfn

Co-authored-by: ppfn <benjamin.desrousseaux@gmx.fr>