Fixed a bug that caused the `isa` parameter written in the `ics` file to be wrong in most cases
The existing code considered `ispar(5)` to contain the value of the `isa` parameter, and this was actually only true for `ibasty`=4. For other bases, whether they have an `isa` parametered stored somewhere else or not, the code wrote the value of `ispar(5)`, which usually contains another parameter of the base.

Now the code writes the value of the `isa` parameter if the base has it, zero otherwise.

I fixed all the reference `ics` files so that they have the proper value for `isa`  parameter. This way, all the tests pass.

fixes #90