added the test `test_c2hh2_long`
This tests is a version of `test_c2hh2` that lasts longer (about 5 minutes (323s) on `graffy-ws2` in release mode : core i5 8th generation).

starting from `c2hh2_long`, which had:
- jtot2 = 10
- n1max = 6
- j2max = 1

made the folloing changes to reach 5 minutes
| action       | num_channels | duration(s) |
|              |          147 |          25 |
| j2max <- 2   |          147 |          22 |
| n1max <- 8   |          243 |          48 |
| jtot2 <- 12  |          243 |          63 |
| n1max <- 10  |          363 |         128 |
| n1max <- 14  |          661 |         323 |

nb: this test belongs to the new `benchmark` testsuite, and is not supposed to be part of the `coverage` test suite, as it doesn't add any code coverage compared to the existing tests in `coverage` test suite. The `benchmark` suite is intended to be used by the extensive continuous integration (see issue #102)
nb: this test was chosen because it uses base 26, which is used to test the growable arrays (see issue #51)