replaced `corpar` common block with `mod_par` module
This one was less straightforward than the others, as each parameter in rpar array has 2 different names depending on whether scattering mode or bound state mode is used. Both modes have different set of real parameters but is it just by luck that there are as many parameters in both modes? We need to replace this with something less hacky...