fixed use of unitialized floating value
This bug existed in both `dsigga` and `dsigh`, which duplicated 99% of their code, making it hard to maintain.

So I refactored `dsigga` and `dsigh` to unify their code. This was a bit tedious though, as it required to upgrade some function with 'implicit none' (which is a good thing anyway)

fixes #137