fixed bug that caused hibridon to not behave properly when wfu files are bigger than 2^31
This was caused by the fact that hibridon used 32 bits integers to store positions in the file relative to the beginning (in `iendwv`, `ipos2`, `ipos3`). To fix this issus, these positions are now stored in 64 bit integers (this will therefore allow files  containig up to 2^63 bytes, which is far beyond our usage)

important note : because `ipos2` and `ipos3` are now 64 bits integers instead of 32 bits integers, the file format of wfu files is now different (the new format reveived the version number 3, as the previous one was 2) from previous versions of hibridon (8 bytes bigger). To avoid any confusion, hibridon will detect and display an approriate error message if it's attempting to load an unsupported (ie different from 3)  wfu file format

took this opportunity to clean up the code:
- converted the common block `cowave` into a module `mod_wave`
- documented some magic numbers related to wfu files
- added utility functions to unify redundant code
- refactored entry `waverd` into a proper subroutine
- removed some unused variables to unclutter the warnings