commit of the progress made in june 2022. These changes were stashed but now it's hard to remember... so here's what I remember :
These changes are part of the aim at converting the base 01 to using param_group instead of scod and ispar, in order to get rid of error prone indices. For example

```
scod(1)='NTERM'
iscod(1) = 1
```
is now replaced with

```
par_nterm = basis_params%create_iparam('NTERM')
par_nterm%set_value(1)
```

However, I beleive that this work is not finished (it builds but arn2 test (a test that uses base01) fails).
- finished ptr1sg and sav1sg so that the code now compiles
- one of the difficulties I remember was to separate  ptr1sg into a proper subroutine (it used to be an entry, which shared local variables with other subroutines). Not sure that this part is finished, as there's still some debug code...
- started to handle param_group in hibridon's intepreter (hinput), but most of it is still disabled to make code compile