removed unwanted warnings about unused dummy arguments
empty subroutines such as `ground` and `bausr` don't use any of their dummy arguments, and this is intentional, as these subroutines act as templates. In order to keep these arguments without causing warnings, this commit introduces a hacky solution to fool the compiler into thinking that these arguments are used.