converted cotrans common block into 2 separate modules as they don't need to be the same data
One of these arrays is specific to base 7 and the other is specific to base 10 so I put this data into their relevant base.
ttrans should be seen as base 07's member data and base 10's member data. We now better see that ttrans is base specific data.

work related to #7