fixed bug introduced recently while refactoring common block costp
This bug wasn't undetected on gfortran builds because of the presence of the compilation flag -fno-automatic on gfortran builds. This flag caused all local variables to be saved between calls, practiclly making these variables global (with a local scope). Saving all local variables is unneccary in most cases; we should use the save modifier only for variables that need it, as it's easier to understand code with non-saved local variables. So, I removed the automatic save of the local variables because:
- it's not needed (the code passes all tests without it)
- it's not a good practice
- now ifort and gfortran build use the same paradigm (no automatic savesd variables)