added support for output files where only the absolute value matters
This is expected to fix the test `caheflux`, which fails on some builds, because the flx file contents differ.

For some reason, the sign of the values in the transformation matrix can vary depending on the build environment (compiler or blas libraries). Unless this comes from a bug (eg uninitialized variable), I assume that both signs are correct, and that this sign variation comes from multiple valid solutions to the same problem.

work related to issue #34 and #245